+++ src/js/index.ts
@@ -91,11 +91,13 @@ export { TouchEventBoundary, withTouchEventBoundary } from './touchevents';
 export {
   ReactNativeTracing,
   ReactNavigationV4Instrumentation,
-  // eslint-disable-next-line deprecation/deprecation
   ReactNavigationV5Instrumentation,
   ReactNavigationInstrumentation,
   ReactNativeNavigationInstrumentation,
   RoutingInstrumentation,
+  reactNativeTracingIntegration,
+  reactNavigationIntegration,
+  reactNativeNavigationIntegration,
   sentryTraceGesture,
   TimeToInitialDisplay,
   TimeToFullDisplay,