+++ src/js/tools/metroconfig.ts
@@ -1,5 +1,6 @@
 import { logger } from '@sentry/utils';
 import type { MetroConfig, MixedOutput, Module, ReadOnlyGraph } from 'metro';
+import type { CustomResolutionContext, CustomResolver, Resolution } from 'metro-resolver';
 import * as process from 'process';
 import { env } from 'process';
 
@@ -7,7 +8,6 @@ import { enableLogger } from './enableLogger';
 import { cleanDefaultBabelTransformerPath, saveDefaultBabelTransformerPath } from './sentryBabelTransformerUtils';
 import { createSentryMetroSerializer, unstable_beforeAssetSerializationPlugin } from './sentryMetroSerializer';
 import type { DefaultConfigOptions } from './vendor/expo/expoconfig';
-
 export * from './sentryMetroSerializer';
 
 enableLogger();
@@ -18,6 +18,11 @@ export interface SentryMetroConfigOptions {
    * @default false
    */
   annotateReactComponents?: boolean;
+  /**
+   * Adds the Sentry replay package for web.
+   * @default true
+   */
+  includeWebReplay?: boolean;
 }
 
 export interface SentryExpoConfigOptions {
@@ -35,7 +40,7 @@ export interface SentryExpoConfigOptions {
  */
 export function withSentryConfig(
   config: MetroConfig,
-  { annotateReactComponents = false }: SentryMetroConfigOptions = {},
+  { annotateReactComponents = false, includeWebReplay = true }: SentryMetroConfigOptions = {},
 ): MetroConfig {
   setSentryMetroDevServerEnvFlag();
 
@@ -45,6 +50,9 @@ export function withSentryConfig(
   newConfig = withSentryFramesCollapsed(newConfig);
   if (annotateReactComponents) {
     newConfig = withSentryBabelTransformer(newConfig);
+  }
+  if (includeWebReplay === false) {
+    newConfig = withSentryResolver(newConfig, includeWebReplay);
   }
 
   return newConfig;
@@ -71,6 +79,10 @@ export function getSentryExpoConfig(
   let newConfig = withSentryFramesCollapsed(config);
   if (options.annotateReactComponents) {
     newConfig = withSentryBabelTransformer(newConfig);
+  }
+
+  if (options.includeWebReplay === false) {
+    newConfig = withSentryResolver(newConfig, options.includeWebReplay);
   }
 
   return newConfig;
@@ -142,6 +154,66 @@ function withSentryDebugId(config: MetroConfig): MetroConfig {
     serializer: {
       ...config.serializer,
       customSerializer,
+    },
+  };
+}
+
+// Based on: https://github.com/facebook/metro/blob/c21daba415ea26511e157f794689caab9abe8236/packages/metro-resolver/src/resolve.js#L86-L91
+type CustomResolverBeforeMetro068 = (
+  context: CustomResolutionContext,
+  realModuleName: string,
+  platform: string | null,
+  moduleName?: string,
+) => Resolution;
+
+/**
+ * Includes `@sentry/replay` packages based on the `includeWebReplay` flag and current bundle `platform`.
+ */
+export function withSentryResolver(config: MetroConfig, includeWebReplay: boolean | undefined): MetroConfig {
+  const originalResolver = config.resolver?.resolveRequest as CustomResolver | CustomResolverBeforeMetro068 | undefined;
+
+  const sentryResolverRequest: CustomResolver = (
+    context: CustomResolutionContext,
+    moduleName: string,
+    platform: string | null,
+    oldMetroModuleName?: string,
+  ) => {
+    if (
+      (includeWebReplay === false ||
+        (includeWebReplay === undefined && (platform === 'android' || platform === 'ios'))) &&
+      (oldMetroModuleName ?? moduleName).includes('@sentry/replay')
+    ) {
+      return { type: 'empty' } as Resolution;
+    }
+    if (originalResolver) {
+      return oldMetroModuleName
+        ? originalResolver(context, moduleName, platform, oldMetroModuleName)
+        : originalResolver(context, moduleName, platform);
+    }
+
+    // Prior 0.68, resolve context.resolveRequest is sentryResolver itself, where on later version it is the default resolver.
+    if (context.resolveRequest === sentryResolverRequest) {
+      // eslint-disable-next-line no-console
+      console.error(
+        `Error: [@sentry/react-native/metro] Can not resolve the defaultResolver on Metro older than 0.68.
+Please follow one of the following options:
+- Include your resolverRequest on your metroconfig.
+- Update your Metro version to 0.68 or higher.
+- Set includeWebReplay as true on your metro config.
+- If you are still facing issues, report the issue at http://www.github.com/getsentry/sentry-react-native/issues`,
+      );
+      // Return required for test.
+      return process.exit(-1);
+    }
+
+    return context.resolveRequest(context, moduleName, platform);
+  };
+
+  return {
+    ...config,
+    resolver: {
+      ...config.resolver,
+      resolveRequest: sentryResolverRequest,
     },
   };
 }